[MIG] Migration to 12.0: date_search_extended [MERGE] date_search_extended relavent changes to coop_account and coop_point_of_sale
