+++ coop_account/models/account_move.py
@@ -7,6 +7,17 @@ class AccountMove(models.Model):
     _inherit = 'account.move'
 
     payment_id = fields.Many2one('account.payment', 'Payment Entry')
+    search_year = fields.Char(
+        string='Year (Search)', compute='_compute_date_search',
+        multi='_date_search', store=True, index=True)
+
+    search_month = fields.Char(
+        string='Month (Search)', compute='_compute_date_search',
+        multi='_date_search', store=True, index=True)
+
+    search_day = fields.Char(
+        string='Day (Search)', compute='_compute_date_search',
+        multi='_date_search', store=True, index=True)
 
     @api.multi
     def unmatch_bankstatement(self):
@@ -55,3 +66,18 @@ class AccountMove(models.Model):
         res = super(AccountMove, self).write(vals)
         self.check_bank_statement_journal()
         return res
+
+    @api.multi
+    @api.depends('date')
+    def _compute_date_search(self):
+        """ Merge from date_search_extended module from version 9
+        remove date_search_extended module from version 12"""
+        for rec in self:
+            if rec.date:
+                rec.search_year = rec.date.strftime('%Y')
+                rec.search_month = rec.date.strftime('%Y-%m')
+                rec.search_day = rec.date.strftime('%Y-%m-%d')
+            else:
+                rec.search_year = False
+                rec.search_month = False
+                rec.search_day = False
