+++ coop_point_of_sale/views/view_pos_session.xml
@@ -11,6 +11,16 @@
                 <field name="week_day"/>
                 <field name="week_number"/>
             </field>
+            <field name="user_id" position="after">
+                <field name="search_year"/>
+                <field name="search_month"/>
+                <field name="search_day"/>
+            </field>
+            <xpath expr="//group" position="inside">
+                <filter string="Year (Search)" name="year" context="{'group_by':'search_year'}"/>
+                <filter string="Month (Search)" name="month" context="{'group_by':'search_month'}"/>
+                <filter string="Day (Search)" name="day" context="{'group_by':'search_day'}"/>
+            </xpath>
         </field>
     </record>
