--- intercoop_addons/date_search_extended/README.rst
@@ -1,16 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
-==========================
-Coop Custom - Date Search
-==========================
-
-This module adds date fields to make date search functions more complete.
-
-Credits
-=======
-
-Contributors
-------------
-
-* Sylvain LE GAL <https://twitter.com/legalsylvain>
