--- intercoop_addons/date_search_extended/models/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import account_move
-from . import account_move_line
-from . import pos_order
-from . import pos_session
