--- intercoop_addons/date_search_extended/models/account_move_line.py
@@ -1,40 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from datetime import datetime
-from openerp import models, fields, api
-
-
-class AccountMoveLine(models.Model):
-    _inherit = 'account.move.line'
-
-    _SEARCH_DATE_FIELD = 'date'
-
-    search_year = fields.Char(
-        string='Year (Search)', compute='_compute_date_search',
-        multi='_date_search', store=True, index=True)
-
-    search_month = fields.Char(
-        string='Month (Search)', compute='_compute_date_search',
-        multi='_date_search', store=True, index=True)
-
-    search_day = fields.Char(
-        string='Day (Search)', compute='_compute_date_search',
-        multi='_date_search', store=True, index=True)
-
-    @api.multi
-    @api.depends('date')
-    def _compute_date_search(self):
-        date_scheme = '%Y-%m-%d'
-        for item in self:
-            if item.date:
-                my_date = datetime.strptime(item.date, date_scheme)
-                item.search_year = my_date.strftime('%Y')
-                item.search_month = my_date.strftime('%Y-%m')
-                item.search_day = my_date.strftime('%Y-%m-%d')
-            else:
-                item.search_year = False
-                item.search_month = False
-                item.search_day = False
