--- intercoop_addons/date_search_extended/views/view_pos_session.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-
-<odoo>
-
-    <record model="ir.ui.view" id="view_pos_session_search">
-        <field name="model">pos.session</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_session_search"/>
-        <field name="arch" type="xml">
-            <field name="user_id" position="after">
-                <field name="search_year"/>
-                <field name="search_month"/>
-                <field name="search_day"/>
-            </field>
-            <xpath expr="//group" position="inside">
-                <filter string="Year (Search)" context="{'group_by':'search_year'}"/>
-                <filter string="Month (Search)" context="{'group_by':'search_month'}"/>
-                <filter string="Day (Search)" context="{'group_by':'search_day'}"/>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
